Factored out a scriban utility
