+++ Sources/Props/SourceGenerator.Build.props
@@ -1,5 +1,7 @@
 <!-- Source Generator project settings -->
 <Project>
+  <Import Project="Shared.Build.props" />
+
   <PropertyGroup>
     <TargetFramework>netstandard2.0</TargetFramework>
     <LangVersion>preview</LangVersion>
